+++ bake_filter.h
@@ -28,7 +28,7 @@ namespace bake
   void filter(
       const Instance*      instances,
       const size_t         num_instances,
-      const unsigned int*  num_samples_per_instance,
+      const size_t*        num_samples_per_instance,
       const AOSamples&     ao_samples,
       const float*         ao_values,
       float**              vertex_ao
